requirements.txt
# FastAPI 웹 프레임워크
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2

# 인증 및 보안
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
cryptography==43.0.1
slowapi==0.1.9

# 데이터베이스
sqlalchemy==2.0.32
pymysql==1.1.1
pymongo==4.8.0
redis==5.0.8
opensearch-py==2.7.0

# Kafka
kafka-python==2.0.2

# HTTP 클라이언트
requests==2.32.3
httpx==0.27.0

# 유틸리티
python-dotenv==1.0.1
pyyaml==6.0.2
tqdm==4.66.5
tenacity==9.0.0
python-dateutil==2.9.0.post0
